Major Project Assignment - Part 1
Part 1 – Description and Basic Functional Prototype:
A. Create a GitHub repository and description page for your software.  Address the who, what,
where, when, how, and why for the software:
Who are the users of the software?
What is the purpose of the software?
Where and when will the software be used?
How does the software work?
Why would anyone want to use the software over existing processes?
B. Create UML class diagrams to describe the relationship between the classes in your software.
Save these UML diagrams as PNG images and post the UML class diagrams, with appropriate
captions, in your GitHub description page.
C. Create a basic functional prototype for the software. The prototype must:
Read/write data from/to a file
Use a graphical user interface (e.g., JavaFX or Swing, unless given permission by the instructor
otherwise)
Include an array or ArrayList data structure
Include an efficient sorting algorithm (e.g., quicksort / mergesort / heapsort) with comments
describing the efficiency,
Use generics and classes with an inheritance structure of your own creation to match your
UML.
Upload this version of the software to your GitHub account via IntelliJ using the information
given in Blackboard.